refactor: router 관련 로직 및 import를 main에서 분리
